requirements.txt
deepface 
flask 
flask-cors
tf-keras
python-dotenv